Update for ITISE 2023
TEMOA-Europe version to be used for runs up to 2050. This TEMOA-Europe version has been used for the work "Analysis of the long-term effects of the restrictions on trade with Russia using the  TEMOA-Europe energy system optimization model" presented at the 9th International conference on Time Series and Forecasting (ITISE 2023) at Gran Canaria, Spain.